+++ source/Util.cpp
@@ -17,7 +17,7 @@ namespace Puny
 	static const int s_cWrapInitial = 0x80;
 	static const u32 s_nU32Max = 0xFFFFFFFF;
 
-	int NAdaptBias(int nDelta, int cPoint, bool fFirstTime)
+	ptrdiff_t NAdaptBias(ptrdiff_t nDelta, ptrdiff_t cPoint, bool fFirstTime)
 	{
 		if (fFirstTime) nDelta = nDelta / s_nDamp;
 		else			nDelta = nDelta / 2;
@@ -207,8 +207,8 @@ PUNYRET PunyretEncode(const char * pCozInput, char * pCozOut, size_t cBMaxOut)
 			}
 		}
 	}
-	int cBasic = pCozDest - pCozOut;
-	int cHandled = pCozDest - pCozOut;
+	ptrdiff_t cBasic = pCozDest - pCozOut;
+	ptrdiff_t cHandled = pCozDest - pCozOut;
 
 	// BB - would like to change this to only add a delimiter when we have extended characters (or another delimiter)
 	if ((pNExt - aNExtSorted) != 0 || fFoundDelimiter)
@@ -224,12 +224,12 @@ PUNYRET PunyretEncode(const char * pCozInput, char * pCozOut, size_t cBMaxOut)
 
 	std::sort(aNExtSorted, pNExt);
 
-	int nDelta = 0;
-	int cWrap = s_cWrapInitial;
-	int nBias = s_nBiasInitial;
+	ptrdiff_t nDelta = 0;
+	ptrdiff_t cWrap = s_cWrapInitial;
+	ptrdiff_t nBias = s_nBiasInitial;
 
 	int * pNWorkMax = pN;
-	int cCodepoint = pNWorkMax - aNWork;
+	ptrdiff_t cCodepoint = pNWorkMax - aNWork;
 
 	int iSorted = 0;
 	while (cHandled < cCodepoint)
@@ -251,10 +251,10 @@ PUNYRET PunyretEncode(const char * pCozInput, char * pCozOut, size_t cBMaxOut)
 			}
 			else if (nC == cWrap)
 			{
-				int nQ = nDelta;
+				ptrdiff_t nQ = nDelta;
 				for (int nK = s_nBase; 1; nK += s_nBase)
 				{
-					int t = EWC::ewcMax(EWC::ewcMin(nK - nBias, s_nTMax), s_nTMin);
+					int t = EWC::ewcMax(EWC::ewcMin(int(nK - nBias), s_nTMax), s_nTMin);
 					if (nQ < t)
 						break;
 
@@ -262,7 +262,7 @@ PUNYRET PunyretEncode(const char * pCozInput, char * pCozOut, size_t cBMaxOut)
 					nQ = (nQ - t) / (s_nBase - t);
 				}
 				
-				*pCozDest++ = ChEncodeDigit(nQ);
+				*pCozDest++ = ChEncodeDigit(u32(nQ));
 		        nBias = NAdaptBias(nDelta, cHandled + 1, cHandled == cBasic);
 				nDelta = 0;
 				++cHandled;
@@ -328,10 +328,10 @@ PUNYRET PunyretDecode(const char * pCozInput, char * pCozOut, size_t cBOutMax)
 	if (!fAllBasic)	// skip the delimiter
 		++pCoz;
 
-	u32 cWrap = s_cWrapInitial; // aka 'n'
+	ptrdiff_t cWrap = s_cWrapInitial; // aka 'n'
 	u32 iPrev;
 	u32 iN = 0;
-	u32 nBias = s_nBiasInitial;
+	ptrdiff_t nBias = s_nBiasInitial;
 	u32 nWeight;
 	u32 nK;
 	u32 t;
@@ -357,7 +357,7 @@ PUNYRET PunyretDecode(const char * pCozInput, char * pCozOut, size_t cBOutMax)
 
 			if (nK <= nBias)				t = s_nTMin;
 			else if (nK >= nBias + s_nTMax) t = s_nTMax;
-			else							t = nK - nBias;
+			else							t = u32(nK - nBias);
 
 			if (nDigit < t)
 				break;
@@ -371,7 +371,7 @@ PUNYRET PunyretDecode(const char * pCozInput, char * pCozOut, size_t cBOutMax)
 
 	    // iN was supposed to wrap around from out+1 to 0, incrementing n each time, so we'll fix that now:
 
-		u32 dWrap = iN / (cNOut + 1);
+		ptrdiff_t dWrap = iN / (cNOut + 1);
 	    if (dWrap > s_nU32Max  - cWrap) 
 			return PUNYRET_Overflow;
 	    cWrap += dWrap;
@@ -382,7 +382,7 @@ PUNYRET PunyretDecode(const char * pCozInput, char * pCozOut, size_t cBOutMax)
 	    if (cNOut >= cBOutMax) 
 			return PUNYRET_OutputTooLong;
 	    memmove(aNOutput + iN + 1, aNOutput + iN, (cNOut - iN) * sizeof(aNOutput[0]));
-		aNOutput[iN++] = cWrap;
+		aNOutput[iN++] = u32(cWrap);
 	}
 
 	aNOutput[cNOut] = 0;
